refactor(common): 패키지명 리팩토링
