+++ src/main/kotlin/racingcar/domain/const/StringRes.kt
@@ -1,0 +1,13 @@
+package racingcar.domain.const
+
+const val UNDER_BAR = "_"
+const val CAR_NAME_FORMAT = "%s : %s"
+const val EMPTY = ""
+const val HOW_MANY_CAR = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."
+const val ERROR_CAR_NAME_OVERFLOW = "자동차 이름은 5자를 넘을 수 없습니다."
+const val HOW_MANY_TURN = "시도할 횟수는 몇 회인가요?"
+const val GAME_RESULT = "실행 결과"
+const val CAR_NUMBER_ERROR = "자동차 대수 입력이 잘못 되었습니다."
+const val TURN_NUMBER_ERROR = "횟수 입력이 잘못 되었습니다."
+const val WINNER_IS = "가 최종 우승했습니다."
+const val WINNER_SEPARATOR = ", "
