+++ src/main/kotlin/racingcar/domain/result/RacingResult.kt
@@ -1,0 +1,5 @@
+package racingcar.domain.result
+
+import racingcar.domain.car.Car
+
+class RacingResult(val cars: List<Car>)
