--- src/main/kotlin/racingcar/racing/result/RacingResult.kt
@@ -1,5 +1,0 @@
-package racingcar.racing.result
-
-import racingcar.racing.car.Car
-
-class RacingResult(val cars: List<Car>)
