--- src/main/kotlin/racingcar/racing/view/ResultView.kt
@@ -1,10 +1,0 @@
-package racingcar.racing.view
-
-class ResultView {
-    fun showResult(racingResult: String, winnerResult: String) {
-        // 레이싱 결과 출력
-        println(racingResult)
-        // 시상식
-        println(winnerResult)
-    }
-}
