+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -1,0 +1,10 @@
+package racingcar.view
+
+class ResultView {
+    fun showResult(racingResult: String, winnerResult: String) {
+        // 레이싱 결과 출력
+        println(racingResult)
+        // 시상식
+        println(winnerResult)
+    }
+}
