+++ src/test/kotlin/racingcar/domain/rule/MockUpRule.kt
@@ -1,0 +1,9 @@
+package racingcar.domain.rule
+
+class MockUpRule : CarMovementRule {
+    var hasBeenCalled = false
+    override val rule = {
+        hasBeenCalled = true
+        true
+    }
+}
