--- src/test/kotlin/racingcar/racing/rule/MockUpRule.kt
@@ -1,9 +1,0 @@
-package racingcar.racing.rule
-
-class MockUpRule : CarMovementRule {
-    var hasBeenCalled = false
-    override val rule = {
-        hasBeenCalled = true
-        true
-    }
-}
